Fix : Renaming test-utils to testutils, updating Makefile with trigger package tests
